fix references broken due to storage refactoring
